requirements.txt
firebase==4.0.1
firebase-admin==6.5.0
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.28
langchain-experimental==0.0.62
langchain-google-firestore==0.3.0
langchain-google-genai==1.0.7
langchain-openai==0.1.10
langchainhub==0.1.20